Rename 'config' package and add package-info to various packages
Renamed the 'config' package to 'model' for better clarity. Additionally, added package-info.java files with @NotNull annotations to multiple packages for enforcing non-null constraints and better code documentation.
