+++ src/main/java/net/thenextlvl/tweaks/command/warp/package-info.java
@@ -1,0 +1,10 @@
+@TypesAreNotNullByDefault
+@FieldsAreNotNullByDefault
+@MethodsReturnNotNullByDefault
+@ParametersAreNotNullByDefault
+package net.thenextlvl.tweaks.command.warp;
+
+import core.annotation.FieldsAreNotNullByDefault;
+import core.annotation.MethodsReturnNotNullByDefault;
+import core.annotation.ParametersAreNotNullByDefault;
+import core.annotation.TypesAreNotNullByDefault;
